stream: add brotli support to `CompressionStream` and `DecompressionStream`
